# 🚨 Critical Performance Fix: Reduce API payload by 99.998% (251MB → 5KB)